+++ src/main/java/com/yhr/smcp/repositories/character/mythicplus/MythicSeasonRepository.java
@@ -12,5 +12,5 @@ import java.util.Optional;
 public interface MythicSeasonRepository extends JpaRepository<MythicSeason, Long> {
     Page<MythicSeason> findByProfileId(Long profileId, Pageable pageable);
 
-    Optional<MythicSeason> findByProfileIdAndKeystoneSeasonId(Long profileId, Integer keystoneSeasonId);
+    Optional<MythicSeason> findByProfileIdAndKeystoneSeasonId(Long profileId, Long keystoneSeasonId);
 }
